% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairml-package.R
\docType{package}
\name{clinicalfair-package}
\alias{clinicalfair}
\alias{clinicalfair-package}
\title{clinicalfair: Algorithmic Fairness for Clinical Prediction Models}
\description{
Evaluates algorithmic fairness of clinical prediction models across
demographic subgroups. Computes standard fairness metrics, generates
disparity plots, and provides threshold-based mitigation strategies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CuiweiG/clinicalfair}
  \item Report bugs at \url{https://github.com/CuiweiG/clinicalfair/issues}
}

}
\author{
\strong{Maintainer}: Cuiwei Gao \email{48gaocuiwei@gmail.com} [copyright holder]

}
\keyword{internal}
