% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration curves by group}
\usage{
plot_calibration(data, n_bins = 10L)
}
\arguments{
\item{data}{A \link{fairness_data} object.}

\item{n_bins}{Number of calibration bins. Default 10.}
}
\value{
A ggplot object.
}
\description{
Assesses whether predicted probabilities match observed event
rates within each protected group.
}
\examples{
\donttest{
data(compas_sim)
fd <- fairness_data(compas_sim$risk_score, compas_sim$recidivism,
                    compas_sim$race)
plot_calibration(fd)
}

}
