% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness-plot.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curves by group}
\usage{
plot_roc(data)
}
\arguments{
\item{data}{A \link{fairness_data} object.}
}
\value{
A ggplot object.
}
\description{
Plot ROC curves by group
}
\examples{
\donttest{
set.seed(42)
fd <- fairness_data(
  predictions = c(runif(100, 0.2, 0.8), runif(100, 0.3, 0.9)),
  labels = c(rbinom(100, 1, 0.3), rbinom(100, 1, 0.5)),
  protected_attr = rep(c("A", "B"), each = 100)
)
plot_roc(fd)
}

}
